Split EnchantmentHelper and EnchantmentShopHelper
