+++ CSharpSourceCode/TOR_Core.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32505.173
+# Visual Studio Version 18
+VisualStudioVersion = 18.9.12120.119 stable
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TOR_Core", "TOR_Core.csproj", "{80942161-02FB-4024-A1C2-22EF33DC8D1A}"
 EndProject
@@ -13,8 +13,8 @@ Global
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|Any CPU.Build.0 = Debug|x64
 		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|x64.ActiveCfg = Debug|x64
 		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Debug|x64.Build.0 = Debug|x64
 		{80942161-02FB-4024-A1C2-22EF33DC8D1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
